+++ example/example_server/lib/src/server/server_state.dart
@@ -1,0 +1,3 @@
+class ServerState {
+  int counter = 0;
+}