+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3-rc.4
+- Depend on serverpod RC
+- [FEAT]: Server context for sharing dependencies across all endpoints and executions
+
 ## 0.0.2-alpha.3
 - Depend on serverpod alpha